Allow webpack helpers to be used from data files too
